+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 * Improved detection of translation with full language code.
 * Improved detection of language code in directory and file name.
 * Improved detection of language code separated by full stop.
+* Added detection for app store metadata files.
 
 1.0
 ---